chore: Move core module to core.main
