+++ core/main/src/androidMain/AndroidManifest.xml
