+++ core/main/src/androidMain/kotlin/eu/kanade/tachiyomi/network/NetworkHelper.kt
