+++ core/main/src/androidMain/kotlin/eu/kanade/tachiyomi/util/system/ToastExtensions.kt
@@ -1,0 +1,37 @@
+package eu.kanade.tachiyomi.util.system
+
+import android.content.Context
+import android.widget.Toast
+import androidx.annotation.StringRes
+import dev.icerock.moko.resources.StringResource
+import yokai.util.lang.getString
+
+/**
+ * Display a toast in this context.
+ *
+ * @param resource the text resource.
+ * @param duration the duration of the toast. Defaults to short.
+ */
+fun Context.toast(@StringRes resource: Int, duration: Int = Toast.LENGTH_SHORT) {
+    Toast.makeText(this, resource, duration).show()
+}
+
+/**
+ * Display a toast in this context.
+ *
+ * @param resource the text resource.
+ * @param duration the duration of the toast. Defaults to short.
+ */
+fun Context.toast(resource: StringResource, duration: Int = Toast.LENGTH_SHORT) {
+    toast(getString(resource), duration)
+}
+
+/**
+ * Display a toast in this context.
+ *
+ * @param text the text to display.
+ * @param duration the duration of the toast. Defaults to short.
+ */
+fun Context.toast(text: String?, duration: Int = Toast.LENGTH_SHORT) {
+    Toast.makeText(this, text.orEmpty(), duration).show()
+}
