+++ core/main/src/androidMain/kotlin/yokai/core/archive/AndroidArchiveInputStream.kt
@@ -1,0 +1,61 @@
+package yokai.core.archive
+
+import java.nio.ByteBuffer
+import kotlin.concurrent.Volatile
+import me.zhanghai.android.libarchive.Archive
+import me.zhanghai.android.libarchive.ArchiveEntry
+import me.zhanghai.android.libarchive.ArchiveException
+
+class AndroidArchiveInputStream(buffer: Long, size: Long) : ArchiveInputStream() {
+    private val lock = Any()
+    @Volatile
+    private var isClosed = false
+
+    private val archive = Archive.readNew()
+
+    init {
+        try {
+            Archive.setCharset(archive, Charsets.UTF_8.name().toByteArray())
+            Archive.readSupportFilterAll(archive)
+            Archive.readSupportFormatAll(archive)
+            Archive.readOpenMemoryUnsafe(archive, buffer, size)
+        } catch (e: ArchiveException) {
+            close()
+            throw e
+        }
+    }
+
+    private val oneByteBuffer = ByteBuffer.allocateDirect(1)
+
+    override fun read(): Int {
+        read(oneByteBuffer)
+        return if (oneByteBuffer.hasRemaining()) oneByteBuffer.get().toUByte().toInt() else -1
+    }
+
+    override fun read(b: ByteArray, off: Int, len: Int): Int {
+        val buffer = ByteBuffer.wrap(b, off, len)
+        read(buffer)
+        return if (buffer.hasRemaining()) buffer.remaining() else -1
+    }
+
+    private fun read(buffer: ByteBuffer) {
+        buffer.clear()
+        Archive.readData(archive, buffer)
+        buffer.flip()
+    }
+
+    override fun close() {
+        synchronized(lock) {
+            if (isClosed) return
+            isClosed = true
+        }
+
+        Archive.readFree(archive)
+    }
+
+    fun getNextEntry() = Archive.readNextHeader(archive).takeUnless { it == 0L }?.let { entry ->
+        val name = ArchiveEntry.pathnameUtf8(entry) ?: ArchiveEntry.pathname(entry)?.decodeToString() ?: return null
+        val isFile = ArchiveEntry.filetype(entry) == ArchiveEntry.AE_IFREG
+        ArchiveEntry(name, isFile)
+    }
+}
