+++ core/main/src/androidMain/kotlin/yokai/core/archive/ArchiveEntry.kt
