+++ core/main/src/androidMain/kotlin/yokai/core/archive/ArchiveInputStream.kt
