+++ core/main/src/commonMain/kotlin/eu/kanade/tachiyomi/network/NetworkPreferences.kt
