--- core/src/androidMain/kotlin/eu/kanade/tachiyomi/util/system/UniFileExtensions.kt
@@ -1,54 +1,0 @@
-package eu.kanade.tachiyomi.util.system
-
-import android.content.Context
-import android.os.Build
-import android.os.FileUtils
-import android.os.ParcelFileDescriptor
-import com.hippo.unifile.UniFile
-import java.io.BufferedOutputStream
-import java.io.File
-import java.nio.channels.SeekableByteChannel
-
-val UniFile.nameWithoutExtension: String?
-    get() = name?.substringBeforeLast('.')
-
-val UniFile.extension: String?
-    get() = name?.replace("${nameWithoutExtension.orEmpty()}.", "")
-
-val UniFile.displayablePath: String
-    get() = filePath ?: uri.toString()
-
-fun UniFile.toTempFile(context: Context): File {
-    val inputStream = context.contentResolver.openInputStream(uri)!!
-    val tempFile =
-        File.createTempFile(
-            nameWithoutExtension.orEmpty(),
-            null,
-        )
-
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-        FileUtils.copy(inputStream, tempFile.outputStream())
-    } else {
-        BufferedOutputStream(tempFile.outputStream()).use { tmpOut ->
-            inputStream.use { input ->
-                val buffer = ByteArray(8192)
-                var count: Int
-                while (input.read(buffer).also { count = it } > 0) {
-                    tmpOut.write(buffer, 0, count)
-                }
-            }
-        }
-    }
-
-    return tempFile
-}
-
-fun UniFile.writeText(string: String, onComplete: () -> Unit = {}) {
-    this.openOutputStream().use {
-        it.write(string.toByteArray())
-        onComplete()
-    }
-}
-
-fun UniFile.openFileDescriptor(context: Context, mode: String): ParcelFileDescriptor =
-    context.contentResolver.openFileDescriptor(uri, mode) ?: error("Failed to open file descriptor: $displayablePath")
