+++ settings.gradle.kts
@@ -31,7 +31,7 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 rootProject.name = "Yokai"
 include(":app")
-include(":core")
+include(":core:main")
 include(":data")
 include(":domain")
 include(":i18n")
